requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
qdrant-client==1.9.1
sentence-transformers==2.7.0
PyMuPDF==1.24.3
groq==0.9.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
